Decompile aiDoPresetAnimation and fix mismatching setup files

+++ src/files/setup/mp_setupmp11.c
@@ -4,7 +4,7 @@
 
 #include "stagesetup.h"
 
-u8 intro[];
+s32 intro[];
 u8 props[];
 struct path paths[];
 struct ailist ailists[];
@@ -81,7 +81,7 @@ u8 props[] = {
 	endprops
 };
 
-u8 intro[] = {
+s32 intro[] = {
 	spawn(0x001c)
 	spawn(0x001d)
 	spawn(0x001e)
